fix: correct item IDs in tutorial quest rewards
The tutorial quest rewards were using incorrect item IDs:
- 'apple' -> 'food_apple'
- 'rubber_ball' -> 'toy_ball'

This caused item rewards to not be added to inventory on quest
completion because the addItem function silently returns the
original inventory when the item ID doesn't exist.

Also added a test to verify item rewards are properly granted
when completing quests.